Move predict-only ccb data out of ccb_label to make slim build
