+++ vowpalwabbit/ccb_label.h
@@ -10,6 +10,8 @@
 #include "label_parser.h"
 #include "v_array.h"
 #include "action_score.h"
+// TODO: This header can be removed once type and explicit_included_actions are removed from the label
+#include "ccb_predict_data.h"
 
 namespace CCB
 {
@@ -23,13 +25,6 @@ struct conditional_contextual_bandit_outcome
   ACTION_SCORE::action_scores probabilities;
 };
 
-enum example_type : uint8_t
-{
-  unset = 0,
-  shared = 1,
-  action = 2,
-  slot = 3
-};
 
 //TODO: Remove the elements that are in predict_data
 // ccb_label.cc will need a major revamp before that can happen
@@ -40,14 +35,6 @@ struct label
   conditional_contextual_bandit_outcome* outcome;
   v_array<uint32_t> explicit_included_actions;
   float weight;
-};
-
-struct predict_data
-{
-  example_type type;
-  v_array<uint32_t> explicit_included_actions;
-
-  void clear() { explicit_included_actions.clear(); }
 };
 
 extern label_parser ccb_label_parser;
